Refactor ChapterEditor into focused hooks and components
